+++ lib/dentaku/ast/function.rb
@@ -1,4 +1,5 @@
 require_relative 'node'
+require_relative 'function_registry'
 
 module Dentaku
   module AST
@@ -12,61 +13,19 @@ module Dentaku
       end
 
       def self.get(name)
-        registry.fetch(function_name(name)) {
-          fail ParseError, "Undefined function #{ name }"
-        }
+        registry.get(name)
       end
 
       def self.register(name, type, implementation)
-        function = Class.new(self) do
-          def self.implementation=(impl)
-            @implementation = impl
-          end
-
-          def self.implementation
-            @implementation
-          end
-
-          def self.type=(type)
-            @type = type
-          end
-
-          def self.type
-            @type
-          end
-
-          def value(context={})
-            args = @args.map { |a| a.value(context) }
-            self.class.implementation.call(*args)
-          end
-
-          def type
-            self.class.type
-          end
-        end
-
-        function_class = name.to_s.capitalize
-        Dentaku::AST.send(:remove_const, function_class) if Dentaku::AST.const_defined?(function_class, false)
-        Dentaku::AST.const_set(function_class, function)
-
-        function.implementation = implementation
-        function.type = type
-
-        registry[function_name(name)] = function
+        registry.register(name, type, implementation)
       end
 
       def self.register_class(name, function_class)
-        registry[function_name(name)] = function_class
-      end
-
-      private
-
-      def self.function_name(name)
-        name.to_s.downcase
+        registry.register_class(name, function_class)
       end
 
       def self.registry
-        @registry ||= {}
+        @registry ||= FunctionRegistry.new
       end
     end
   end