+++ lib/dentaku/parser.rb
@@ -7,8 +7,9 @@ module Dentaku
     def initialize(tokens, options={})
       @input      = tokens.dup
       @output     = []
-      @operations = options.fetch(:operations, [])
-      @arities    = options.fetch(:arities, [])
+      @operations        = options.fetch(:operations, [])
+      @arities           = options.fetch(:arities, [])
+      @function_registry = options.fetch(:function_registry, nil)
     end
 
     def get_args(count)
@@ -25,6 +26,9 @@ module Dentaku
 
       while token = input.shift
         case token.category
+        when :datetime
+          output.push AST::DateTime.new(token)
+
         when :numeric
           output.push AST::Numeric.new(token)
 
@@ -211,6 +215,8 @@ module Dentaku
         pow:      AST::Exponentiation,
         negate:   AST::Negation,
         mod:      AST::Modulo,
+        bitor:    AST::BitwiseOr,
+        bitand:   AST::BitwiseAnd,
 
         lt:       AST::LessThan,
         gt:       AST::GreaterThan,
@@ -225,7 +231,11 @@ module Dentaku
     end
 
     def function(token)
-      Dentaku::AST::Function.get(token.value)
+      function_registry.get(token.value)
+    end
+
+    def function_registry
+      @function_registry ||= Dentaku::AST::FunctionRegistry.new
     end
   end
 end